+++ example/namedtracer/main.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"log"
 
-	"go.opentelemetry.io/otel/api/correlation"
+	"go.opentelemetry.io/otel/api/baggage"
 	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/example/namedtracer/foo"
@@ -62,7 +62,7 @@ func main() {
 	tracer := tp.Tracer("example/namedtracer/main")
 	ctx := context.Background()
 
-	ctx = correlation.NewContext(ctx,
+	ctx = baggage.NewContext(ctx,
 		fooKey.String("foo1"),
 		barKey.String("bar1"),
 	)